Wallet Controllers

+++ src/main/java/com/proriberaapp/ribera/services/client/WalletService.java
@@ -4,6 +4,8 @@ import com.proriberaapp.ribera.Domain.entities.WalletEntity;
 import com.proriberaapp.ribera.Domain.entities.WalletTransactionEntity;
 import reactor.core.publisher.Mono;
 
+import java.math.BigDecimal;
+
 public interface WalletService {
 
 
@@ -12,10 +14,11 @@ public interface WalletService {
   Mono<WalletEntity> createWalletUsuario(Integer userId, Integer currencyId);
   Mono<WalletEntity> createWalletPromoter(Integer promoterId, Integer currencyId);
 
-  Mono<WalletTransactionEntity> makeTransfer (Integer walletIdOrigin, Integer walletIdDestiny ,String emailDestiny, String nameDestiny, Double amount);
-  Mono<WalletTransactionEntity> makeWithdrawal(Integer walletId, Integer transactioncatid, Double amount);
-  Mono<WalletTransactionEntity> makeDeposit(Integer walletId, Integer transactioncatid, Double amount);
-  Mono<WalletTransactionEntity> makePayment(Integer walletId, Integer transactioncatid, Double amount);
-  Mono<WalletTransactionEntity> makeRecharge(Integer walletId, Integer transactioncatid, Double amount);
+  Mono<WalletTransactionEntity> makeTransfer (Integer walletIdOrigin, Integer walletIdDestiny ,String emailDestiny, String nameDestiny, BigDecimal amount);
+  Mono<WalletTransactionEntity> makeWithdrawal(Integer walletId, Integer transactioncatid, BigDecimal amount);
+  Mono<WalletTransactionEntity> makeDeposit(Integer walletId, Integer transactioncatid, BigDecimal amount);
+  Mono<WalletTransactionEntity> makePayment(Integer walletId, Integer transactioncatid, BigDecimal amount);
+  Mono<WalletTransactionEntity> makeRecharge(Integer walletId, Integer transactioncatid, BigDecimal amount);
+
 
 }
